+++ pkg/tsdb/cloudwatch/query_builder_test.go
@@ -41,110 +41,6 @@ func TestCloudWatchQueryBuilder(t *testing.T) {
 				So(*res[0].Id, ShouldEqual, "id1_____0")
 				So(*res[1].Id, ShouldEqual, "id1_____1")
 			})
-
-			Convey("and query expression will be used if it was set in query editor", func() {
-				query := &CloudWatchQuery{
-					RefId:      "A",
-					Region:     "us-east-1",
-					Expression: "SEARCH(someexpression)",
-					Statistics: []*string{aws.String("Average")},
-					Period:     300,
-					Id:         "id1",
-					Identifier: "id1",
-				}
-
-				res, err := e.buildMetricDataQueries(query)
-				So(err, ShouldBeNil)
-				So(len(res), ShouldEqual, 1)
-				So(*res[0].Expression, ShouldEqual, "SEARCH(someexpression)")
-			})
-
-			Convey("and metric stat", func() {
-				Convey("will be used when expression is not set in the client and not more than one dimension value is used for on specific key and a star is not being used as dimension value", func() {
-					dimensions := make(map[string][]string)
-					dimensions["InstanceId"] = []string{"i-12345678"}
-					query := &CloudWatchQuery{
-						RefId:      "A",
-						Region:     "us-east-1",
-						Expression: "",
-						Statistics: []*string{aws.String("Average")},
-						Period:     300,
-						Id:         "id1",
-						Identifier: "id1",
-						Dimensions: dimensions,
-					}
-
-					res, err := e.buildMetricDataQueries(query)
-					So(err, ShouldBeNil)
-					So(len(res), ShouldEqual, 1)
-					So(res[0].Expression, ShouldBeNil)
-					So(res[0].MetricStat, ShouldNotBeNil)
-				})
-
-				Convey("will not used when expression is set in the client", func() {
-					dimensions := make(map[string][]string)
-					dimensions["InstanceId"] = []string{"i-12345678"}
-					query := &CloudWatchQuery{
-						RefId:      "A",
-						Region:     "us-east-1",
-						Expression: "SEARCH(",
-						Statistics: []*string{aws.String("Average")},
-						Period:     300,
-						Id:         "id1",
-						Identifier: "id1",
-						Dimensions: dimensions,
-					}
-
-					res, err := e.buildMetricDataQueries(query)
-					So(err, ShouldBeNil)
-					So(len(res), ShouldEqual, 1)
-					So(res[0].Expression, ShouldNotBeNil)
-					So(res[0].MetricStat, ShouldBeNil)
-				})
-
-				Convey("will be not used when a dimension key has more than one value", func() {
-					dimensions := make(map[string][]string)
-					dimensions["InstanceId"] = []string{"i-12345678", "i-34562312"}
-					query := &CloudWatchQuery{
-						RefId:      "A",
-						Region:     "us-east-1",
-						Expression: "",
-						Statistics: []*string{aws.String("Average")},
-						Period:     300,
-						Id:         "id1",
-						Identifier: "id1",
-						Dimensions: dimensions,
-					}
-
-					res, err := e.buildMetricDataQueries(query)
-					So(err, ShouldBeNil)
-					So(len(res), ShouldEqual, 1)
-					So(res[0].Expression, ShouldNotBeNil)
-					So(res[0].MetricStat, ShouldBeNil)
-				})
-
-				Convey("will be not used when a dimension key has a value that is equals to *", func() {
-					dimensions := make(map[string][]string)
-					dimensions["InstanceId"] = []string{"i-12345678", "*"}
-					query := &CloudWatchQuery{
-						RefId:      "A",
-						Region:     "us-east-1",
-						Expression: "",
-						Statistics: []*string{aws.String("Average")},
-						Period:     300,
-						Id:         "id1",
-						Identifier: "id1",
-						Dimensions: dimensions,
-					}
-
-					res, err := e.buildMetricDataQueries(query)
-					So(err, ShouldBeNil)
-					So(len(res), ShouldEqual, 1)
-					So(res[0].Expression, ShouldNotBeNil)
-					So(res[0].MetricStat, ShouldBeNil)
-				})
-			})
-
 		})
 
 		Convey("buildSearchExpression", func() {
