--- pkg/tsdb/cloudwatch/types.go
@@ -1,31 +1,0 @@
-package cloudwatch
-
-import "fmt"
-
-type CloudWatchQuery struct {
-	RefId              string
-	Region             string
-	Id                 string
-	Namespace          string
-	MetricName         string
-	Statistics         []*string
-	QueryType          string
-	Expression         string
-	ReturnData         bool
-	Dimensions         map[string][]string
-	ExtendedStatistics []*string
-	Period             int
-	Alias              string
-	Identifier         string
-	HighResolution     bool
-	SearchExpressions  []string
-}
-
-type queryBuilderError struct {
-	err   error
-	RefID string
-}
-
-func (e *queryBuilderError) Error() string {
-	return fmt.Sprintf("Error parsing query %s, %s", e.RefID, e.err)
-}
